add enhanced SDK autoconfig for exporters
